moved some types around to fix pydantic models
