+++ mypy-baseline.txt
@@ -902,8 +902,6 @@ src/inmanta/protocol/methods.py:0: error: Missing type parameters for generic ty
 src/inmanta/protocol/methods.py:0: note: Use "-> None" if function does not return a value
 src/inmanta/protocol/methods.py:0: note: Use "-> None" if function does not return a value
 src/inmanta/protocol/methods.py:0: note: Use "-> None" if function does not return a value
-src/inmanta/protocol/methods_v2.py:0: error: Cannot determine type of "RVID_OPTS"  [has-type]
-src/inmanta/protocol/methods_v2.py:0: error: Cannot determine type of "RVID_OPTS"  [has-type]
 src/inmanta/protocol/methods_v2.py:0: error: Missing return statement  [empty-body]
 src/inmanta/protocol/methods_v2.py:0: error: Missing return statement  [empty-body]
 src/inmanta/protocol/methods_v2.py:0: error: Missing return statement  [empty-body]
