+++ src/inmanta/types.py
@@ -19,17 +19,17 @@ Contact: code@inmanta.com
 # This file defines named type definition for the Inmanta code base
 
 import builtins
+import datetime
 import uuid
 from collections.abc import Coroutine, Generator, Mapping, Sequence
-from datetime import datetime
-from typing import TYPE_CHECKING, Any, Callable, NewType, Optional, Union
+from typing import TYPE_CHECKING, Any, Callable, ClassVar, NewType, Optional, Union
+from inmanta.stable_api import stable_api
 
 import pydantic
 import typing_inspect
 
 if TYPE_CHECKING:
     # Include imports from other modules here and use the quoted annotation in the definition to prevent import loops
-    from inmanta.data.model import BaseModel  # noqa: F401
     from inmanta.protocol.common import ReturnValue  # noqa: F401
 
 
@@ -41,6 +41,43 @@ else:
     import _typeshed
 
     DataclassProtocol = _typeshed.DataclassInstance
+
+
+def api_boundary_datetime_normalizer(value: datetime.datetime) -> datetime.datetime:
+    if value.tzinfo is None:
+        return value.replace(tzinfo=datetime.timezone.utc)
+    else:
+        return value
+
+
+@stable_api
+class DateTimeNormalizerModel(pydantic.BaseModel):
+    """
+    A model that normalizes all datetime values to be timezone aware. Assumes that all naive timestamps represent UTC times.
+    """
+
+    @pydantic.field_validator("*", mode="after")
+    @classmethod
+    def validator_timezone_aware_timestamps(cls: type, value: object) -> object:
+        """
+        Ensure that all datetime times are timezone aware.
+        """
+        if isinstance(value, datetime.datetime):
+            return api_boundary_datetime_normalizer(value)
+        else:
+            return value
+
+
+@stable_api
+class BaseModel(DateTimeNormalizerModel):
+    """
+    Base class for all data objects in Inmanta
+    """
+
+    # Populate models with the value property of enums, rather than the raw enum.
+    # This is useful to serialise model.dict() later
+    model_config: ClassVar[pydantic.ConfigDict] = pydantic.ConfigDict(use_enum_values=True)
+
 
 # kept for backwards compatibility
 StrictNonIntBool = pydantic.StrictBool
@@ -65,7 +102,7 @@ def issubclass(sub: type, super: Union[type, tuple[type, ...]]) -> bool:
     return builtins.issubclass(sub, super)
 
 
-PrimitiveTypes = Optional[uuid.UUID | bool | int | float | datetime | str]
+PrimitiveTypes = Optional[uuid.UUID | bool | int | float | datetime.datetime | str]
 type SimpleTypes = BaseModel | PrimitiveTypes
 
 JsonType = dict[str, Any]
@@ -74,6 +111,7 @@ type StrictJson = dict[str, StrictJson] | list[StrictJson] | str | int | float |
 
 type StrMapping[T] = Mapping[str, T] | Mapping[ResourceIdStr, T] | Mapping[ResourceVersionIdStr, T]
 
+# TODO: this PR changed mapping values from SimpleTypes to ArgumentTypes. Why? Was it just inaccurate?
 type SinglePageTypes = SimpleTypes | StrMapping[ArgumentTypes]
 # only simple types allowed within list args, not dicts or lists.
 # Typed as Sequence for necessity (covariance), though runtime checks and method overloads require list in practice.
@@ -86,10 +124,10 @@ type PageableTypes = Sequence[SimpleTypes]
 type ArgumentTypes = SinglePageTypes | PageableTypes
 type ReturnTypes = SinglePageTypes | PageableTypes
 
-type MethodReturn = ReturnTypes | ReturnValue[ReturnTypes]
+type MethodReturn = ReturnTypes | "ReturnValue[ReturnTypes]"
 type MethodType = Callable[..., MethodReturn]
 
-type Apireturn = int | ReturnTupple | ReturnValue[ReturnTypes] | ReturnValue[None] | ReturnTypes
+type Apireturn = int | ReturnTupple | "ReturnValue[ReturnTypes]" | "ReturnValue[None]" | ReturnTypes
 type Warnings = Optional[list[str]]
 type HandlerType = Callable[..., AsyncioCoroutine[Apireturn]]
